Initial commit
This i s a place to rebuild [squasch][0] by moving parts over rather than
by editing the original (which has a *lot* of things we don't need
anymore).

My hope is that this will make it easier to keep the types checked, the code
working, and the tests green while I make changes to squasch.

Also, shrg is a *lot* easier to type than squasch