Simplify types in field parsers
This comes with a loss of specificity in that all of the parsers now
have the same return type even though they will each only have one
or two output types (e.g StringParser and NumberParser now have
the same return type).

However, the code that uses the parsers deals with fields generically,
so we end up with collections of different typed things anyways. I think
the loss of specificity is worth the gain in ease-of-use, and it points
to how the code will really be called.