Separate the load and parse steps of data loading
This commit separates the load and parse steps of CSV rows because we can't
tell from the individual CSV rows whether they're well formed or not. In
the updated, multi-row syntax, "required" fields may be ommitted to
indicate that the row is part of an extended record. This should be checked
when the rows are transformed into records, but not when the CSV is
being loaded.

By saving the contents of each row and the location in the CSV file,
we can still raise useful errors if the row is indeed malformed, but can
defer parsing and validating until we have the information to do it.