Change function in row module to reflect deferred parsing
Since I've split the loading process into a load step (which saves
each field's location and source value) and parsing (after the records
for a case have been gathered and we can check, for example, which
required fields are missing), the main function the row module was
misleadingly named. This commit renames it for consistency.