Add case parsing
This commit adds a module with functions for grouping and parsing
input rows into data collected for individual cases.

It might more properly be called "tokenizing", but I think that might
be confusing to anyone who's not in the specific frame of mind of a
compiler constructor.

It also adds the dependency for `hypothesis`, a generative-testing
framework, and ignores the development files it generates while
running tests.