Remove dev tools dependency file
These are particular to me, and are installed globally in my dev
environments (that ended up being the least fussy way to manage them,
and they change slowly enough that isolating them for each project isn't
very beneficial). Since the dependencies in this file are either
optional or recorded in the `setup.py` file, the file is being removed.